Backend/inference_server/requirements.txt
flask==3.0.2
tensorflow==2.17.0
mediapipe==0.10.14
numpy==1.26.4
opencv-python-headless==4.10.0.84
mysql-connector-python==9.1.0
PyJWT==2.9.0
bcrypt==4.2.0